[docs] Add "See Also" blocks for `torch.t` and `torch.transpose`
**Summary**
PR 63177 added "see also" cross-references to `torch.t` and
`torch.transpose`. However, it did so in the form of simple links
preceded by the phrase "See also". A comment on the aforementioned PR
pointed out the RST has special syntax to render a "See Also" box that
is more noticeable. This commit converts the "see also" sentences for
`torch.t` and `torch.transpose` to boxes.

**Test Plan**
Continuous integration.

**Fixes**
This commit fixes #56267.

ghstack-source-id: d921f05806a29ea1421d31680eabec3df44e3f5b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63461