[PyTorch] Use SmallVector for FunctionSchema Arguments
Heap allocation profiling seemed to show that this path
allocates on the heap a bunch, but the typical function has a small
number of arguments, with 1 being very common and the count of
functions with each number of arguments decreasing after that. I
picked 3 as a guess for what would provide a good tradeoff between
wasted inline space and not allocating on the heap.

Differential Revision: [D30816470](https://our.internmc.facebook.com/intern/diff/D30816470/)

ghstack-source-id: 137603695
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64691